Removed compiler warning for Xcode 4.5 64-bit
Former-commit-id: c8b80c57e43fa5dc62508ddd5117555550834210

+++ cocos2d/CCParticleBatchNode.m
@@ -124,7 +124,7 @@
 
 -(NSString*) description
 {
-	return [NSString stringWithFormat:@"<%@ = %p | Tag = %i>", [self class], self, tag_ ];
+	return [NSString stringWithFormat:@"<%@ = %p | Tag = %ld>", [self class], self, (long)tag_ ];
 }
 
 -(void)dealloc
